Implement applyManifestsSDK
