+++ pkg/app/pipedv1/plugin/kubernetes/deployment/apply.go
@@ -17,11 +17,25 @@ package deployment
 import (
 	"context"
 	"errors"
+	"time"
 
 	"github.com/pipe-cd/pipecd/pkg/app/pipedv1/plugin/kubernetes/provider"
 	"github.com/pipe-cd/pipecd/pkg/plugin/logpersister"
+	"github.com/pipe-cd/pipecd/pkg/plugin/sdk"
 )
 
+type applier interface {
+	// ApplyManifest does applying the given manifest.
+	ApplyManifest(ctx context.Context, manifest provider.Manifest) error
+	// CreateManifest does creating resource from given manifest.
+	CreateManifest(ctx context.Context, manifest provider.Manifest) error
+	// ReplaceManifest does replacing resource from given manifest.
+	ReplaceManifest(ctx context.Context, manifest provider.Manifest) error
+	// ForceReplaceManifest does force replacing resource from given manifest.
+	ForceReplaceManifest(ctx context.Context, manifest provider.Manifest) error
+}
+
+// TODO: rewrite applyManifests with sdk.StageLogPersister
 func applyManifests(ctx context.Context, applier applier, manifests []provider.Manifest, namespace string, lp logpersister.StageLogPersister) error {
 	if namespace == "" {
 		lp.Infof("Start applying %d manifests", len(manifests))
@@ -75,5 +89,19 @@ func applyManifests(ctx context.Context, applier applier, manifests []provider.M
 
 	}
 	lp.Successf("Successfully applied %d manifests", len(manifests))
+	return nil
+}
+
+// TODO: remove applyManifestsSDK and originalLogPersister after rewriting applyManifests
+func applyManifestsSDK(ctx context.Context, applier applier, manifests []provider.Manifest, namespace string, lp sdk.StageLogPersister) error {
+	originalLogPersister := &originalLogPersister{lp}
+	return applyManifests(ctx, applier, manifests, namespace, originalLogPersister)
+}
+
+type originalLogPersister struct {
+	sdk.StageLogPersister
+}
+
+func (lp *originalLogPersister) Complete(timeout time.Duration) error {
 	return nil
 }
